requirements.txt
openai
python-dotenv
requests
pydantic>=2.0.0 